+++ ebl/tests/factories/archaeology.py
@@ -1,5 +1,5 @@
 from ebl.fragmentarium.domain.archaeology import Archaeology, ExcavationNumber
-from ebl.fragmentarium.domain.iso_date import DateRange, DateWithNotes
+from ebl.fragmentarium.domain.date_range import DateRange, DateWithNotes
 from ebl.fragmentarium.domain.findspot import BuildingType, ExcavationPlan, Findspot
 from ebl.tests.factories.bibliography import ReferenceFactory
 from ebl.tests.factories.collections import TupleFactory
